Extract near-best structural search out of the inference solver (#256).
